+++ managers/slack_manager.py
